+++ stellar-multitenant/internal/httphandler/tenants_handler_test.go
@@ -346,6 +346,7 @@ func Test_TenantHandler_Post(t *testing.T) {
 			"auth_user_mfa_codes",
 			"auth_user_password_reset",
 			"auth_users",
+			"bridge_integration",
 			"circle_client_config",
 			"circle_recipients",
 			"circle_transfer_requests",